feat: migrate embeddings from MiniLM to OpenAI text-embedding-3-small
Move collection schemas from site/search/ to baker/typesense/ so the
OpenAI API key can be set directly without runtime indirection.
site/search/typesenseCollections.ts now only exports document interfaces.
